backend/requirements.txt
requests
python-dotenv
fastapi
uvicorn
google-cloud-storage